Mv ContentRepository to models.models
