+++ ansible_galaxy/installed_repository_db.py
@@ -3,8 +3,7 @@ import os
 
 from ansible_galaxy import matchers
 from ansible_galaxy import installed_namespaces_db
-from ansible_galaxy.models.content_repository import ContentRepository
-from ansible_galaxy.models.models import ContentSpec
+from ansible_galaxy.models.models import ContentRepository, ContentSpec
 
 log = logging.getLogger(__name__)
